Moved CheckBirthdayWorker.Launcher to BootBroadcastReceiver
