+++ app/src/main/java/ch/epfl/reminday/background/BirthdayWorkerFactory.kt
@@ -8,6 +8,9 @@ import ch.epfl.reminday.data.birthday.BirthdayDao
 import ch.epfl.reminday.format.date.DateFormatter
 import java.util.*
 
+/**
+ * A subclass of [WorkerFactory] that can create [CheckBirthdaysWorker]s.
+ */
 class BirthdayWorkerFactory(
     private val birthdayDao: BirthdayDao,
     private val locale: Locale,
@@ -18,10 +21,11 @@ class BirthdayWorkerFactory(
         workerClassName: String,
         workerParameters: WorkerParameters
     ): ListenableWorker? {
-        val formatter = DateFormatter.shortFormatter(locale)
         return when (workerClassName) {
-            CheckBirthdaysWorker::class.java.name ->
+            CheckBirthdaysWorker::class.java.name -> {
+                val formatter = DateFormatter.shortFormatter(locale)
                 CheckBirthdaysWorker(birthdayDao, formatter, appContext, workerParameters)
+            }
             // Return null, so that the base class can delegate to the default WorkerFactory
             else ->
                 null
